Basic Folder setup, HTML & SCSS
I have created the folder structure, the general HTML structure and have started styling the page for mobile view.

I have added buttons to the Navigation Page which change upon activation. I would like to get a transition working, however it does not appear to be working at this moment in time.