Navigation, Hero and Projects
I have changed the Navigation so that it sits above everything else on the page, and isn't affected by overlap from other elements.

I have changed the Hero Image to be a background image with text as the main element.

I have changed some styling for the projects section and converted it to display as a flex-box.